requirements.txt
# Core dependencies
requests>=2.31.0
urllib3>=2.0.0
beautifulsoup4>=4.12.0
colorama>=0.4.6
rich>=13.0.0
click>=8.1.0
pycryptodome>=3.19.0
cryptography>=41.0.0
scapy>=2.5.0
python-nmap>=0.7.1
selenium>=4.15.0
lxml>=4.9.0
fake-useragent>=1.4.0
dnspython>=2.4.0
paramiko>=3.3.0
psutil>=5.9.0
tqdm>=4.66.0
pyfiglet>=0.8.0
termcolor>=2.3.0
nltk>=3.8.1
pyyaml>=6.0.1

# Hash cracking and cryptography
bcrypt>=4.0.1
argon2-cffi>=21.3.0
scrypt>=0.8.18
pbkdf2>=1.3
PyJWT>=2.8.0

# Tor and proxy support
PySocks>=1.7.1

# Additional security libraries
passlib[bcrypt]>=1.7.4

# Data analysis and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Async and networking
asyncio>=3.4.3
aiohttp>=3.8.0
websockets>=11.0.0
# uvloop>=0.17.0  # Not compatible with Windows

# Web framework for dashboard and API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0

# Real-time analysis
sseclient-py>=1.7.0

# Plugin system
importlib-metadata>=6.0.0
watchdog>=3.0.0

# Auto-completion
# readline>=6.2.0  # Not compatible with Windows

# Additional security and reconnaissance tools
python-whois>=0.8.0

# Additional security tools
python-whois>=0.8.0
shodan>=1.30.0
censys>=2.2.0
virustotal-api>=1.1.11

# Machine learning and AI
tensorflow>=2.13.0
torch>=2.0.0
transformers>=4.30.0
openai>=1.0.0
anthropic>=0.7.0
langchain>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
tiktoken>=0.5.0
torch-audio>=2.0.0
torchvision>=0.15.0

# Data validation
pydantic>=2.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.11.0

# Monitoring and logging
prometheus-client>=0.17.0
structlog>=23.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Cloud and deployment
docker>=6.1.0
kubernetes>=27.2.0
boto3>=1.28.0
azure-identity>=1.14.0
google-cloud-storage>=2.10.0

# Additional utilities
python-dotenv>=1.0.0
click-plugins>=1.1.1
tabulate>=0.9.0
watchdog>=3.0.0
